+++ isy-persistence/src/main/java/de/bund/bva/pliscommon/persistence/usertype/AbstractImmutableStringUserType.java
@@ -22,7 +22,7 @@ import java.sql.SQLException;
 import java.sql.Types;
 
 import org.hibernate.HibernateException;
-import org.hibernate.engine.spi.SessionImplementor;
+import org.hibernate.engine.spi.SharedSessionContractImplementor;
 import org.hibernate.usertype.UserType;
 
 /**
@@ -45,8 +45,8 @@ public abstract class AbstractImmutableStringUserType extends AbstractImmutableU
     /**
      * {@inheritDoc}
      */
-    public Object nullSafeGet(ResultSet rs, String[] names, SessionImplementor sessionImplementor,
-        Object owner) throws HibernateException, SQLException {
+    public Object nullSafeGet(ResultSet rs, String[] names, SharedSessionContractImplementor
+        sharedSessionContractImplementor, Object owner) throws HibernateException, SQLException {
         String value = rs.getString(names[0]);
         if (rs.wasNull()) {
             return null;
@@ -57,8 +57,8 @@ public abstract class AbstractImmutableStringUserType extends AbstractImmutableU
     /**
      * {@inheritDoc}
      */
-    public void nullSafeSet(PreparedStatement st, Object value, int index,
-        SessionImplementor sessionImplementor) throws HibernateException, SQLException {
+    public void nullSafeSet(PreparedStatement st, Object value, int index, SharedSessionContractImplementor
+        sharedSessionContractImplementor) throws HibernateException, SQLException {
         if (value == null) {
             st.setNull(index, SQL_TYPE);
         } else {