+++ isy-task/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.4.0
+- `IFS-325`: Ungenutze Abhaengigkeit auf `isy-ueberwachung` entfernt
 - `IFS-468`: Update auf dozer 6.5.0 / Dozer aus logback.xml entfernt, wenn es nicht verwendet wird
 
 # 1.3.0