+++ isy-task/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.4.0
+- `IFS-325`: Ungenutze Abhaengigkeit auf `isy-ueberwachung` entfernt
+- `IFS-468`: Update auf dozer 6.5.0 / Dozer aus logback.xml entfernt, wenn es nicht verwendet wird
+
 # 1.3.0
 - `IFS-262`: `isyfact-masterpom` deprecated (Abschaffung mit IsyFact 2.0), `isyfact-masterpom-lib` aufgelöst, Bibliotheken benutzen `isyfact-standards` als Parent-POM